Python/requirements.txt
python-dotenv
requests
uuid
aiohttp
solana